+++ src/base/type/traits.h
@@ -1,21 +1,34 @@
 #ifndef TYPE_TRAITS
 #define TYPE_TRAITS
 
-#include <istream>
+#include <functional>
 #include <optional>
-#include <ostream>
-#include <string>
+#include <type_traits>
 
 namespace Type
 {
 
-template <typename T, typename Enable = void>
-struct isoptional : std::false_type
+template <typename, typename = void>
+struct is_optional : std::false_type
 {};
 
 template <typename T>
-struct isoptional<std::optional<T>> : std::true_type
+struct is_optional<std::optional<T>> : std::true_type
 {};
+
+template <typename T> concept is_optional_v = is_optional<T>::value;
+
+template <typename, typename = void>
+struct is_reference_wrapper : std::false_type
+{};
+
+template <typename T>
+struct is_reference_wrapper<std::reference_wrapper<T>> :
+    std::true_type
+{};
+
+template <typename T>
+concept is_reference_wrapper_v = is_reference_wrapper<T>::value;
 
 }