+++ src/chart/generator/marker.h
@@ -31,7 +31,6 @@ public:
 	    const Data::SeriesList &mainAxisList,
 	    const Data::SeriesList &subAxisList,
 	    const Data::MultiIndex &index,
-	    MarkerPosition pos,
 	    bool needMarkerInfo);
 
 	::Anim::Interpolated<ColorBase> colorBase;
@@ -62,12 +61,12 @@ public:
 	Id sizeId;
 
 	MarkerIndex idx;
-	MarkerPosition pos;
+	MarkerPosition pos{};
 
 	struct MarkerIndexPosition
 	{
 		MarkerIndex idx;
-		MarkerPosition pos;
+		MarkerPosition pos{};
 
 		friend bool operator==(const MarkerIndexPosition &lhs,
 		    const MarkerIndexPosition &rhs)