+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## [Unreleased]
 
-### Changed
+### Fixed
 
 - Removed 'min' align property from the API which equivalent with the 'none'.
+- Markers are the same even if new record added.
+
+### Changed
+
+- Changed MarkerId to be a string instead of a number.
+
 
 ## [0.14.0] - 2024-10-03